+++ src/Amadeus/Client.php
@@ -22,16 +22,11 @@
 
 namespace Amadeus;
 
+use Amadeus\Client\Base;
 use Amadeus\Client\Exception;
 use Amadeus\Client\Params;
-use Amadeus\Client\RequestCreator\RequestCreatorInterface;
 use Amadeus\Client\RequestOptions;
-use Amadeus\Client\ResponseHandler\ResponseHandlerInterface;
 use Amadeus\Client\Result;
-use Amadeus\Client\Session\Handler\HandlerFactory;
-use Amadeus\Client\RequestCreator\Factory as RequestCreatorFactory;
-use Amadeus\Client\Session\Handler\HandlerInterface;
-use Amadeus\Client\ResponseHandler\Base as ResponseHandlerBase;
 
 /**
  * Amadeus Web Service Client.
@@ -42,7 +37,7 @@ use Amadeus\Client\ResponseHandler\Base as ResponseHandlerBase;
  * @package Amadeus
  * @author Dieter Devlieghere <dieter.devlieghere@benelux.amadeus.com>
  */
-class Client
+class Client extends Base
 {
     /**
      * Amadeus SOAP header version 1
@@ -62,7 +57,7 @@ class Client
      *
      * @var string
      */
-    const VERSION = "1.2.0-dev";
+    const VERSION = "1.3.0-dev";
 
     /**
      * An identifier string for the library (to be used in Received From entries)
@@ -72,44 +67,9 @@ class Client
     const RECEIVED_FROM_IDENTIFIER = "amabnl-amadeus-ws-client";
 
     /**
-     * Session Handler will be sending all the messages and handling all session-related things.
-     *
-     * @var HandlerInterface
-     */
-    protected $sessionHandler;
-
-    /**
-     * Request Creator is will create the correct message structure to send to the SOAP server.
-     *
-     * @var RequestCreatorInterface
-     */
-    protected $requestCreator;
-
-    /**
-     * Response Handler will check the received response for errors.
-     *
-     * @var ResponseHandlerInterface
-     */
-    protected $responseHandler;
-
-    /**
-     * Authentication parameters
-     *
-     * @var Params\AuthParams
-     */
-    protected $authParams;
-
-    /**
      * @var string
      */
     protected $lastMessage;
-
-    /**
-     * Whether to return the response of a message as XML as well as PHP object
-     *
-     * @var bool
-     */
-    protected $returnResultXml;
 
     /**
      * Set the session as stateful (true) or stateless (false)
@@ -201,47 +161,19 @@ class Client
      *
      * @param Params $params
      */
-    public function __construct($params)
+    public function __construct(Params $params)
     {
-        if ($params->authParams instanceof Params\AuthParams) {
-            $this->authParams = $params->authParams;
-            if (isset($params->sessionHandlerParams) &&
-                $params->sessionHandlerParams instanceof Params\SessionHandlerParams
-            ) {
-                $params->sessionHandlerParams->authParams = $this->authParams;
-            }
-        } elseif (isset($params->sessionHandlerParams) &&
-            $params->sessionHandlerParams->authParams instanceof Params\AuthParams
-        ) {
-            //@deprecated - Provide backwards compatibility with old authparams structure.
-            //Github Issue 40 - retrieve AuthParams from sessionhandlerparams if not generally available
-            $this->authParams = $params->sessionHandlerParams->authParams;
-        }
-
-        $this->sessionHandler = $this->loadSessionHandler(
-            $params->sessionHandler,
-            $params->sessionHandlerParams
+        $this->loadClientParams(
+            $params,
+            self::RECEIVED_FROM_IDENTIFIER,
+            self::VERSION
         );
-
-        $this->requestCreator = $this->loadRequestCreator(
-            $params->requestCreator,
-            $params->requestCreatorParams,
-            self::RECEIVED_FROM_IDENTIFIER."-".self::VERSION,
-            $this->sessionHandler->getOriginatorOffice(),
-            $this->sessionHandler->getMessagesAndVersions()
-        );
-
-        $this->responseHandler = $this->loadResponseHandler(
-            $params->responseHandler
-        );
-
-        $this->returnResultXml = $params->returnXml;
     }
 
     /**
      * Authenticate.
      *
-     * Parameters were provided at construction time (sessionhandlerparams)
+     * Authentication Parameters were provided at construction time (authParams)
      *
      * @return Result
      * @throws Exception
@@ -327,11 +259,6 @@ class Client
      * PNR_RetrieveAndDisplay - Retrieve an Amadeus PNR by record locator including extra info
      *
      * This extra info is info you cannot see in the regular PNR, like Offers.
-     *
-     * By default, the result will be the PNR_RetrieveAndDisplayReply XML as string.
-     * That way you can easily parse the PNR's contents with XPath.
-     *
-     * Set $messageOptions['asString'] = FALSE to get the response as a PHP object.
      *
      * https://webservices.amadeus.com/extranet/viewService.do?id=1922&flavourId=1&menuId=functional
      *
@@ -1141,73 +1068,5 @@ class Client
         }
 
         return $options;
-    }
-
-    /**
-     * Load the session handler
-     *
-     * Either load the provided session handler or create one depending on incoming parameters.
-     *
-     * @param HandlerInterface|null $sessionHandler
-     * @param Params\SessionHandlerParams|null $params
-     * @return HandlerInterface
-     */
-    protected function loadSessionHandler($sessionHandler, $params)
-    {
-        if ($sessionHandler instanceof HandlerInterface) {
-            $newSessionHandler = $sessionHandler;
-        } else {
-            $newSessionHandler = HandlerFactory::createHandler($params);
-        }
-
-        return $newSessionHandler;
-    }
-
-    /**
-     * Load a request creator
-     *
-     * A request creator is responsible for generating the correct request to send.
-     *
-     * @param RequestCreatorInterface|null $requestCreator
-     * @param Params\RequestCreatorParams $params
-     * @param string $libIdentifier Library identifier & version string (for Received From)
-     * @param string $originatorOffice The Office we are signed in with.
-     * @param array $mesVer Messages & Versions array of active messages in the WSDL
-     * @return RequestCreatorInterface
-     * @throws \RuntimeException
-     */
-    protected function loadRequestCreator($requestCreator, $params, $libIdentifier, $originatorOffice, $mesVer)
-    {
-        if ($requestCreator instanceof RequestCreatorInterface) {
-            $newRequestCreator = $requestCreator;
-        } else {
-            $params->originatorOfficeId = $originatorOffice;
-            $params->messagesAndVersions = $mesVer;
-
-            $newRequestCreator = RequestCreatorFactory::createRequestCreator(
-                $params,
-                $libIdentifier
-            );
-        }
-
-        return $newRequestCreator;
-    }
-
-    /**
-     * Load a response handler
-     *
-     * @param ResponseHandlerInterface|null $responseHandler
-     * @return ResponseHandlerInterface
-     * @throws \RuntimeException
-     */
-    protected function loadResponseHandler($responseHandler)
-    {
-        if ($responseHandler instanceof ResponseHandlerInterface) {
-            $newResponseHandler = $responseHandler;
-        } else {
-            $newResponseHandler = new ResponseHandlerBase();
-        }
-
-        return $newResponseHandler;
     }
 }